IGIVEUPWITHTHIS
I am having a mental breakdown so I will eat an apple or something to calm down before I work with this garbage again. My brain is fried and I have no more will to continue with this. What's even sadder is that this is one of my smallest issues. Fuck.